# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/auth/decode/": {
            "post": {
                "description": "Description of the api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Role Create Api Router",
                "operationId": "AuthDecodeController",
                "parameters": [
                    {
                        "description": "Add Role",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.RoleCreated"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Roles",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Role"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/encode/": {
            "post": {
                "description": "Description of the api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Role Create Api Router",
                "operationId": "AuthEncodeController",
                "parameters": [
                    {
                        "description": "Add Role",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.UserCreated"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Roles",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.AuthUser"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/generate-token/{id}/": {
            "get": {
                "description": "Description of the api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Role Create Api Router",
                "operationId": "AuthGenerateTokenController",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Roles",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.AuthUser"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/recovery-email/": {
            "post": {
                "description": "Description of the api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Role Create Api Router",
                "operationId": "AuthRecoveryEmailController",
                "parameters": [
                    {
                        "description": "Add Role",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.UserRecoveryEmail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Roles",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/recovery/": {
            "post": {
                "description": "Description of the api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Role Create Api Router",
                "operationId": "AuthRecoveryController",
                "parameters": [
                    {
                        "description": "Add Role",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.UserRecovery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Roles",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/signin/": {
            "post": {
                "description": "Description of the api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Role Create Api Router",
                "operationId": "AuthSigninController",
                "parameters": [
                    {
                        "description": "Add Role",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.AutCredentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Roles",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.AuthUser"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/signup/": {
            "post": {
                "description": "Description of the api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Role Create Api Router",
                "operationId": "AuthSignupController",
                "parameters": [
                    {
                        "description": "Add Role",
                        "name": "auth",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.UserCreated"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Roles",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.AuthUser"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/user/": {
            "get": {
                "description": "Description of the api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Role Create Api Router",
                "operationId": "AuthUserController",
                "responses": {
                    "200": {
                        "description": "Show Roles",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.AuthUser"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/verify/": {
            "get": {
                "description": "Description of the api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Role Create Api Router",
                "operationId": "AuthVerifyController",
                "responses": {
                    "200": {
                        "description": "Show Roles",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.AuthUser"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "libs.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.AutCredentials": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "models.AuthUser": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "type": "object",
                    "$ref": "#/definitions/models.User"
                }
            }
        },
...
        "models.Role": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                }
            }
        },
        "models.RoleCreated": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "age": {
                    "type": "string",
                    "example": "2006-01-02T00:00:00Z"
                },
                "avatar": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fk_role_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.UserCreated": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "age": {
                    "type": "string",
                    "example": "2006-01-02T00:00:00Z"
                },
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fk_role_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "models.UserRecovery": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "models.UserRecoveryEmail": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}